backend/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.4.0
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.6
email-validator==2.1.1
python-slugify==8.0.1
requests==2.32.3
aiofiles==23.2.1
